Update generator to include baseurl.
One of the breaking errors that I fixed in my install of this great repo was that all links generated by the ruby script didn't take into account the baseurl specified in the config file.

This ruby file corrects the error, although in the paths it might add one '/' too many. This hasn't caused me any issues, but should be considered.